Use a basic extension for CSR, real extensions during signing only
From https://github.com/openssl/openssl/issues/22966
and
https://superuser.com/questions/1778937/the-authoritykeyidentifier-does-not-get-accepted
we discover that the authorityKeyIdentifier is not needed during CSR
generation.  We also discovered that generally everything from the CSR
is discarded during signing anyway.

So we use a basic extension for CSR (using the useless
subjectKeyIdentifier) and then use the real extension during signing
so that it will have whatever we want.